change to project owner approve devices
